removed shared folder, moved Header to Navpages, removed extraneous style sheet
