+++ NavPages/Feed.js
@@ -1,12 +1,14 @@
 import React from 'react';
-import {Text, View} from 'react-native';
+import { Text, View, ScrollView } from 'react-native';
+import Header from './Header';
 
-function About() {
+function Feed() {
   return (
-    <View style={{flex: 1, justifyContent: 'center', alignItems: 'center'}}>
+    <ScrollView>
+      <Header />
       <Text>Feed Page!</Text>
-    </View>
+    </ScrollView>
   );
 }
 
-export default About;
+export default Feed;
